Add shared expat support.
tDOM is distributed with a bundled version of expat (2.0.1).
This is not acceptable in the realm of Debian packaging.
Therefore, this patchs adds support for shared expat version (libexpat1)
as provided by the Debian collection of packages. Once patched, one may
use the "--with-expat" configure flag to point to a shared expat library.
In addition, it renders tDOM indepedent from being either built against
expat 1.95.* or expat 2.* as it offers appropriate preprocessor conditionals
(this applies to both, the bundled and shared scenario).

Based on prior work found at svn://bamboo.nes.ru/debian/tcl/tdom/trunk and
http://autoconf-archive.cryp.to/ax_lib_expat.m4