Extended asHTML with -breakLines and -onlyContents
-breakLines outputs "\n>" instead of ">" for the opening tags of elements.
This makes it less likely that there will be excessively long lines in the
output which break HTML emails - lines longer than 998 chars can be wrapped in
a syntactically invalid way.

-onlyContents omits the outermost open and close tag, returning only the
serialized child nodes (like innerHTML property in a browser).  This allows
HTML fragments to be generated without having to process the result to strip
off the dummy document element wrapper.